+++ src/pgproto/backend/pgproc.rs
@@ -62,7 +62,7 @@ pub fn proc_pg_bind(args: BindArgs) -> PgResult<()> {
         encoding_format: output_format,
     } = args;
 
-    backend::bind(id, stmt_name, portal_name, params, output_format)
+    backend::bind(id, stmt_name, portal_name, params, output_format, vec![])
 }
 
 #[proc]